Introducing NPM to manage dependencies
Introduced the npm package.json file to manage dependencies.
In this case, it's used to download `wrench` as a dev dependency.
This means that typing `npm install` will make sure it is present
before running the other tools.

The goal is to make use of npm to manage other dependencies as well
as run tests, launch the build, etc ...