Recursive unload of classes and templates
Allow classes and templates to be unloaded together with their parents.
In order not to unload core AT stuff, users have to provide a closure that
restricts unloading to their code base, based on the classpath of the
related entity.

Mostly scratching, the same mecanism should be applied to CSS unloading
too, and code cleaned on some parts (like the true callback, and no
tests).