displayReady called too early
When using html templates or when a child template is preloaded,
displayReady of a parent template might get called when the DOM is not yet ready.

The initial fix introduced a performance problem as all preloaded context were considered as differed.
This commit moves the differed logic in the template's init where the information about context ready is more reliable

This fixes #52