fix #303 checkValue error when no bindings
the initial fix for #303 introduces a regression when the widget value is not bound to the datamodel.
If you go from an invalid text a valid value the latter is lost.
moreover if you go from a valid value to an invalid text the widget's style is that of the helptext.