feat #335 configurable number of PhantomJS instances in Attester
Makes it possible to increase (locally, per-user) no. of PhantomJS
instances in Attester to parallelize the execution of the test suite
when launched via 'npm run-script attester' (and 'npm test').

Type in the console e.g.
 npm config set ariatemplates:phantomjsInstances 4
to change the default to 4.

Set it to ~= your processor's noOfCores * noOfThreads * 2
for max performance. E.g. on quad-core dual-threaded processor,
set it to 4*2*2 = 16.

Keep in mind that each thread consumes ~= 30-50 MB of RAM.

This commit requires attester 1.0.2.

Close #335. Close #337.