feat #80 bindable width and height for containers
Adds possibility to bind width and height for some of container-based
widgets (Div, Dialog, Fieldset, TabPanel). The bound size is a subject to
min/max size constraints for all of those widgets.

This commit lifts the requirement for widgets to have width/height = -1
for min/max width/height to work. So this is backward-incompatible change
in some way, but the behavior of width & maxwidth etc. when provided
together was undefined earlier, so it should not have relied upon by anyone.

Apart from the bindable width/height tests, this commit brings also
a test case on static height constraints of the Dialog widget and their
behavior when the contents od the Dialog change due to a refresh.

Close #80. Close #347.