fix #364 File names with multiple dots in atmapreader grunt task
There was a discrepency between the atmapreader grunt task and
aria.core.DownloadMgr.resolveURL.

When removing the extension, the grunt task was looking for the
first dot, whereas the DownloadMgr was looking for the last dot,
so that there was a problem for files containing multiple
dots (such as .tpl.txt, .tpl.css).

This commit fixes the atmapreader grunt task to match the behavior
of the DownloadMgr.