feat #218 better stack trace in Firebug and Safari devtools
This adds support for `Function.displayName` to have better stack
traces in debugger in Firebug and Safari dev tools for the functions
in `$prototype` of classes loaded via `Aria.classDefinition`.

Chrome and IE9+ do not support `displayName` but they have a good
built-in inferring of names of anonymous functions.

IE8- do not support `displayName` nor infer the name by itself.

Close #218. Close #384.