fix #394 tplScript unnecessarily extending Template
If a template had a CSS template with script in its dependencies, then
a circular dependency error might have happened due to that. This commit
removes the the dependency of a tplScript on aria.templates.Template
which could lead to that situation.

Additionally, the $constructor and $destructor functions are removed
at Aria.tplScriptDefinition call time to prevent direct instantiation of
tplScripts (but the user can still specify $constructor and $destructor
in their template script and they will be invoked).

Close #395.