fix #401 HashManager works correctly in IE7 even after a domain change
In order to provide support for back and forward in hash-based navigation in Internet Explorer 7, class aria.utils.HashManager creates a hidden iframe to open and close each time the history has to be populated with an entry.

When the document.domain is changed, IE7 denies access to some global variables of the iframe, unless the document.domain is updated inside the iframe, too. In order to notify that document.domain has been explicly set, variable Aria.domain has to be set to the same value.