refactor global style split & removal
1) Removed page-global CSS from 'widgets.GlobalStyle' and extracted
some of it into 'templates.LegacyGeneralStyle', which is automatically
loaded for backward compatibility in templates when AriaSkin is available.

2) "aria.widgets.GlobalStyle" is loaded only when widgets are present.

3) Removed skin from bootstrap.js. Added skin in attester config instead.

4) In AriaSkinBeans: renamed "GeneralCfg" to "PageGeneralCfg"; removed some
deprecated items from it; created "WidgetGeneralCfg" being subset of
"PageGeneralCfg" to be able to separate font settings for widgets from
font settings for the page in general. This followed some changes in
skin normalization.