[feat] #432 demo attribute for test cases
demoMode can now be set on a testcase object (testObj.demoMode = true) in order to disable the global timeout and to not dispose the template test case at the end of the test.
It can be usefull to keep the ui of a template test case in order to perform manual tests or debug.

To use it in aria tester, just add demo=true in the hash parameters of the url.