fix #520 Accept empty url in navigation manager cache
The navigation manager of the page engine was rejecting empty strings from the cache in which url-to-pageId map was stored.
This implied that, when navigating back to a page whose url is an empty string, the page engine was unable to infer the pageId and display the correct page.

Close #520