feat #518 Addition of storage settings in the page engine site configuration
Navigation managers used both for hash and for history-based navigations in the page engine rely on local storage browser capabilities to store the associations between urls and pageId's. The saved data is needed in order to identify the right page after hard page reload or when hitting back after an external navigation.
Before this code change, the storage could not be configured. Now it can be deactivated, provided with a specific suffix for local storage key and with an expiration delay for its entries.