Fix and simplify page url resolution in the default BasePageProvider
The previous logic fails if the resolution of a path's URL does not end with the path itself

e.g. if  aria.core.DownloadMgr.resolveURL("foo.bar") returns "http://whatever/the/path?path=foo.bar&param=value"

if the DownloadMgr has been customized (thanks to the updateUrlMap(...) function) to be redirected to a servlet with additional parameters (here the 'param=value' appended to the url).

That's why I suggest this simplification including the removal of the variable 'this._basePageUrl' made useless.