fix #573 Style tag reload for html:Template
Calling the getCSSClassNames method of the template context in the initWidget of the html:Template widget was causing a further style tag reload, with a significant performance degradation.

Storing the classnames while writing the markup allows for an optimization of style tags reloads, which is done through the CSSMgr stop/resume mechanism.