fix #588 explain why a frame cannot be loaded
Improve the logging in case a frame cannot be loaded and give it a bit more time to load.
If the iframe doesn't load correctly you might get exceptions like

````
'eval' is undefined
Could not complete the operation due to error 80020101.
````

This also overrides the mechanism for handling errors in `TemplateTestCase` to make sure that the template is disposed in case of errors.
It should make #587 less important, or useless, or an eye-candy while testing.