feat #599 Adding a label.color state property.
The color of the label of most input widgets can now be defined
as a skin property, depending on the state of the widget.

In the atskin skin, the label of text input widgets and of the
select widget now show the disabled state with a different
color than the normal state. Moreover, the appearance of the
MultiSelect (especially for the focused state) is now more
consistent with other text input widgets.

Close #599