feat #668 Test packaged code
With the new version of attester is now possible to test the packaged version of Aria Templates.
This is the code that is ultimately delivered to developers, so it make more sense to test this instead of the development code.

This pull request allows also to test files generated at build time that are not part of the source code.

Close #668