fix #675 Dialog disappears after resize
When the viewport was resized significantly, so that dialogOldPosition.left
> newViewport.width, the Dialog disappeared and never appeared again.

The problem was introduced in e6405a1b79c707f2b4090875ea2e2673f29d5c71 by
the check if reference position is in viewport before displaying the
popup. However this check is valid only in reaction to scrolling events.

The check is now performed only in the onmousewheel event listener.