feat #681: Adding support for the Hebrew language.
This commit adds resource files for the Hebrew language.

Additionally, as, in Hebrew, using substring to deduce the short day
name from the full day name does not work (because all days start
with the same letter), this commit adds the ability to define a
custom short name for days and months. If there is no custom short
name in a resource file, the default mechanism with substring is used.

Close #681