fix #721 changed the way comments are detected in the parser
Instead of checking that there is no ':' in front of the occurrence of a //. It is checked that there are some specific characters or a whitespace.
This fixes the problem with the base64 urls which may contain // somewhere but still allows users to comment their code.

fixed failing test after changing comment detection