fix #838 prefer `updateRootMap` instead of `rootFolderPath` in NodeJS
This commit adds `DownloadMgr.updateRootMap` call in the AT loader for
NodeJS, so that the user can change `Aria.rootFolderPath` to a folder
of his custom files, without worrying of breaking the loading of AT
internal files (`aria.*` classpath).

This is a backward-compatible commit, a prerequisite for #805 which will
change the default `rootFolderPath`.

Close #838.