feat #805 change `rootFolderPath` in NodeJS
This commit changes the `Aria.rootFolderPath` to point to the project's
folder (i.e. the one containing `node_modules`) instead of the AT
installation folder, when the framework is loaded from NodeJS using
`require('ariatemplates')`.

It doesn't make much sense to point to a folder inside
`node_modules/ariatemplates` as a root path, since we won't expect the
user to create any files manually there. Hence it's enough to create the
`aria.*` entry in the URL map, instead of altering `rootFolderPath`.

Note that this commit is backward-incompatible if someone used sth like
this in their NodeJS code:

    folder = Aria.rootFolderPath + "../../../"

    i.e. any relative reference using the `rootFolderPath`.

Close #805.