feat #102 add `timeout` for `RequestMgr.sendJsonRequest`/`submitJsonRequest`
It's been always possible to pass a `timeout` to the calls of
    `aria.core.IO.asyncRequest`
    `aria.core.IO.asyncFormSubmit`
but not to the calls of
    `aria.modules.RequestMgr.submitJsonRequest`
    `aria.modules.RequestMgr.sendJsonRequest`

The latter only had the `aria.core.IO.defaultTimeout` applied.

This commit makes it possible to pass `timeout` as the key of the request
parameter object (first parameter) of them.

Close #102. Close #913.