Always register a binding to section processing
If section processing is not bound to any data model, a spinner overlay created using setProcessingIndicator on the SectionWrapper does not follow it as one of its parent is moved and not closed if the section is deleted. Thus if the user does not specify any binding the overlay remains at its initial place forever.

Binding it to a dummy container allows both of the shortcomings to be solved: spinner follows its parent section and is closed upon section removal.