doc Properties wrongly considered static in APIDoc
The documentation of the code is generated from JSDoc comments, which are parsed with an external tool. It is possible to document let's say detached properties, and to explicitely set their names with a special tag in the JSDoc. However, the tool in question uses this name to determine if a property is attached to an instance, a prototype or is static.

Therefore I prefixed the last part of the names with the proper container name when applicable:

- "prototype" for prototype members
- "instance" for instance members
- nothing for the others, which will be detected as statics (excepts for Aria members which are all considered to be in its prototype)

I took the opportunity to improve the documentation of the Modifiers class, explaining that modifier methods are not actually directly accessible.

Closes #947