fix #980 Correct MultiAutoComplete initialization when bound data is not empty
When freeText option was set to false, items already present in the data model to which the value of the widget is bound were not taken into account.
This commit reverts a fix done for another bug within commit 41d50bfe63ce93bc8d3116d28aeeb753d136b9dd: when freeText is equal to false, type some text for which there is no suggestion. Then blur. Then click inside the field and type something for which suggestions are returned. Pressing ENTER would cause all items to be duplicated.