feat Implemented navigation for the MultiAutoComplete.
In this commit come:

- implementation
- tests
- some refactoring

According to the specifications, it is now possible to use the left and right arrow keys to navigate throughout the whole widget, including currently inserted options. It is also possible to use the tab key to jump form the inserted options list to the text input field, and then from this field to the next focusable element in the page (outside the widget).