feat #1015 Check Git commit message during Travis build
This commit enables checking if commit message of the latest
non-merge Git commit is of good length (not too short, not too long)
and if it adheres to certain team's conventions.

To quickly verify the commit message locally, after the commit, the
developers can use the following commands in their shell:
    grunt git           # (if they have Grunt installed globally)
or
    npm run grunt-git   # local Grunt installation is enough

This commit also provides a commit-time Git hook that can be installed
using this one-liner in the root folder of AT:

    npm run install-hooks

Once the hook is installed locally, if for some reason a developer wants
to bypass the hook, he can commit using the `--no-verify` (`-n`) flag.

Close #1015.