Fixed the behavior on highlighting an option with the click.
Before, anytime a click occured, the text input field was given the focus back, which is a wrong behavior in case of highlighting (navigation wouldn't be possible then, and anyway there is some link between the highlithing state and the focus).

Fixed the test accordingly: anytime an event was simulated, a click event was sent to the text input field, forcing its focus and thus falling into the same issue as described above (and anyway that didn't make sense to do so.)

Besides that a few typo corrections and simplified some code.