feat #1034 Disable method events by default
Backward-incompatible

This commit changes the default value of `this._enableMethodEvents` to
`false` in the `ModuleCtrl` class. The implications of this change are the
new default behaviors when calling any module's public method, as follows:

- no interceptors are added, the methods are called directly
- no events `methodCallBegin`, `methodCallEnd`, `methodCallback` are
  raised
- RefreshMgr is not automatically paused for the span of the call to
  the module controller's method

This commit should bring some performance gains.

In case the developer wants to continue having the old behaviors, he
should set `this._enableMethodEvents = true` in the constructor of his
module controller, *before* calling the parent constructor.

This commit also removes the calls setting the mentioned variable to
false, since they're not needed anymore.

Close #1034.