fix #1033 Removing wrong lines in the MultiAutoComplete.
This commit removes two lines which set a new value in
the data model when pressing tab if the field is empty
and not positioned at the end of the options (i.e. when
leaving the edit mode through tab if the field is empty).

There is no reason to set the value at that time as the
value has not changed, and this can lead to the onchange
handler being called later when the value has not changed.