Auto-escaping now takes into account the case insensitivity of modifiers.
Modifiers are all defined with a lower case name in the framework, but the invokation of one of them can be done by specifying a name with any case: this name gets converted to lower case just before retrieving the actual modifier.

This means that users can write the name of the modifiers with any case, and the automatic escaping feature didn't handle that, thinking that escaping was not controlled by the user when the modifier was used with a case different that the hard-coded one.