feat set/removeDataset methods
This commit introduces two methods to manipulate dataset attributes in HTML element:

* `aria.utils.Html.setDataset`: it allows to specify an object containing key-value pairs of dataset attributes. It uses the keys to add the right attribute, and the values to assign the correct values to them.
* `aria.utils.Html.removeDataset`: it allows to specify an object whose keys will be used to remove dataset attributes.

This feature is compatible with the implementation of `aria.utils.Html.buildAttributeList`, which does not convert camelcase attributes to their dashed counterpart. However, this lack of processing is deprecated (#499). Backward compatible code will have to be removed also from the code added by this commit.