fix #1095 test.aria.widgets.form.autocomplete.samevalue.AutoComplete
test.aria.widgets.form.autocomplete.samevalue.AutoComplete started failing with
commit 1d02f12b58f5d4f84773f2d3605fa544eb096e93.
However, doing the test scenario manually or with the robot works fine.
As a consequence, this pull request changes the test to use the robot instead of Syn.
(Syn puts the cursor at a wrong position when simulating a click on the field).