feat #1106 Include built AT version in npm package
As discussed in #548, it seems that it can be useful to serve minified
version of AT in our npm package along with the src as we're doing now.

This commit effectively adds the contents of `build/target/production`
folder to the npm-published package, and makes the build run automatically
before actually publishing AT to npm via `npm publish` (or `npm pack`).

This increases our npm package's size from ~1.0 MB to ~1.5 MB.
For simplicity, skin files copied at built time to the target folder
are also included in the package (tradeoff simplicity for package size)i.

For applications depending on ariatemplates in their package.json,
they can expose `node_modules/ariatemplates/build/target/production/` as
one of the root folders of their webserver, and they can then include
minified AT in the page via

```
  <script src="/aria/ariatemplates-1.x.x.js"></script>
  <script src="/aria/css/atskin-1.x.x.js"></script>
```

Due to the fact that `npm install` runs also the `prepublish` script,
things were split between `prepublish` and `ci` scripts (the latter being
invoked instead of `npm test` on Travis) in order not to run Grunt/JSHint
twice on Travis.

Close #1106.