fix #1108 Fixing popupWidth property for MAC
This commit aims to fix the `popupWidth` configuration property for the MultiAutoComplete widget.
When the `popupWidth` is smaller than the field width, then it is ognored and set to the field width.

This behaviour is required for other `DropDownListTrait`-based widgets, but not for the MultiAutoComplete. This commit makes so that the `popupWidth` is always applied.