feat #1119 inject page CSS to iframe via TemplateTestCase/FrameATLoader
In TemplateTestCase it's been possible to run a template test in an iframe
and set the `<iframe>`'s CSS itself, but it was not possible to inject the
CSS to the `<body>` of an iframe, and since CSS Templates can't modify global
CSS scope of the window, there way no way to alter global CSS for the
test.

Now it's possible to inject an arbitrary CSS text to the iframe which will
be appended as `<style>` node, as a child of `<head>`.

This fixes the failure of the test
test/aria/widgets/container/dialog/resize/test3/DialogOnResizeTestCase

Close #1119.