fix #1139 DomTestCase fail in Firefox
There's a 0.5px (in this particular scenario) discrepancy in calculations in
Dom.getGeometry method which stems from a quirk in some browsers for
inline containers, for certain font face (Arial, Times New Roman) with
line-height:normal which is a CSS default.

For inline container, it's top position with respect to the parent depends
on the line-height value. For line-height normal, for most of the fonts,
that distance is zero. However for some fonts in some browsers (Firefox,
Chrome abover certain threshold) the computer value of line-height makes
this top position being > 0 and some computations fail later due to that.

This commit just changes the default font for the test. The test started
failing in 1.5.3 after change of the skin from atdefskin to atskin which
removed setting of a default font.

Proper workaround for the issue would require substantial changes to
Dom.getGeometry().

See https://bugzilla.mozilla.org/show_bug.cgi?id=1014738