fix #1162 Fixing a memory leak in SynEvents.js
This commit fixes a memory leak with IE8 in SynEvents.
It is sufficient for aria.utils.SynEvents to be loaded
for the leak to happen.
The issue can be reproduced with attester. As soon as
several tests in the test suite extend (directly or
indirectly) aria.jsunit.TemplateTestCase, because this
class depends on aria.utils.SynEvents, the leak is
visible.
When SynEvents is loaded, this class performs some tests
on the browser to know how it behaves. For this purpose
some DOM elements are created and some event listeners
are attached to them but those listeners were never removed.

This commit removes the leak by removing those event
listeners when they are no longer needed.