fix #1180 Limit call stack in IE
When the `loadFile` method of singleton `aria.core.DownloadMgr` is called, if the file is already available in the AT cache the callback is called synchronously. A big height for the dependency tree of a class leads to a growth of the function call stack which can transcend the limit imposed by some browsers, which leads to an error.

This commit adds a limitation the number of nested synchronous calls to the `loadFile` method in Internet Explorer (64bit), setting a timeout to break javascript execution when the limit is exceeded.