refactor cleanup dead code after #1057
Commit d46d7dd5a938591d6fd8b05622ec4b709e50f389 removed certain
functionalities regarding the usage of `postHeader` but some related
code was not removed properly.

This commit cleans this up by removing the framework code that was
effectively dead as of 1.6.1 (passing a param that was not taken into
account).

The removed test was passing because the cfgbean had `$restricted : false`
hence JSON validator was accepting even the keys removed from the cfgbean
in the earlier commit.

Close #1227.