feat #1229 watch task for rebuilding bootstrap code
This commit adds a watch task (runnable via `npm run watch` or `grunt
watch`) which watches all the files in `src` folder and runs
the bootstrap build whenever any of them changes.

This is meant for developer's convenience (re-running tests in the
browser in the development cycle, via `test/test.htm` page)
until JIRA-403 is implemented.

Since 1.6.1, running tests locally requires a bootstrap build
which was added with the migration to noder-js (commit
7743a70).

Close #1229.