feat #1253 Correct interpretation of dates with localized months
Localized months were not recognized as valid inputs of the `aria.utils.Date.interpret` method.
This means that whenever the user selected a date in a DatePicker with output pattern containing "MMM" or "MMMM" (meaning with the localized short or full month), and then just changing, for example, the year, the entry could not be recognized.

This commit adds this feature. Also, the special case 04DEC/+13 now supports localized dates and the presence of spaces between day, month and year.

Close #1253