feat #1272 removes at-noder-converter from the build
This commit removes the at-noder-converter visitor from the build of
Aria Templates (as source files in the repository are now using the
new syntax).

This commit also changes the configuration of jshint to detect code
which would still use the old syntax (in case of future merges).
For this purpose, it makes sure the Aria global variable is no longer
used (except in resource definitions and skins where it is not possible
to use require to get the equivalent object).

close #1272