feat #1273 Removing the need for a build step after changing files
Before the migration to noder-js, it was possible to run Aria Templates
directly from its src folder.

With the migration to noder-js, this was not possible any more for two
reasons:

1) files in the src folder were using the old syntax (with $dependencies)
and were converted at build time to the new syntax (with require)

-> now, starting with commit b0ae188305c0fa16fe7d458803580da551a1cb94,
files in the repository are using the new syntax and it is no longer
needed to convert them at build time.

2) the entry point file of Aria Templates needs to be generated (as it
is built on top of noder-js, overriding one of its modules) and was
generated in the build/target/bootstrap folder.

-> this commit is changing the location where the entry point file is
generated. It is now generated in the src foilder so that it is possible
to fully execute Aria Templates directly from its src folder.

The generation of files in the src folder is now part of the full build
and can also be run independently with: grunt src

The following files are generated in the src folder:
src/aria/bootstrap.js
src/aria/bootstrap-node.js
src/aria/noderError/*.js

After running "grunt src", it is possible to use the src folder directly
to serve the Aria Templates files, and then changing most of the framework
files can be done without requiring a build step. Running again "grunt src"
is needed after changing noder-js itself or files in the src/noder-modules
folder.

Close #1273