feat #1237 clickbuster implementation
a new touch event has been created (safetap) in order to handle the tap gesture without any side effects due to the ghost click, generated by the browsers in mobile devices.
If a mousedown, click or touchend event is detected after a safetap on the same area, it is ignored since considered as a ghost click: preventDefault() and stopPropagation() are called.
In this way, if the DOM changes after the safetap event is handled, the browser's ghost click will not have effects.

reference:
http://ariatemplates.com/blog/2014/05/ghost-clicks-in-mobile-browsers/