feat #1280: Selenium Java Robot layer
This commit contains a new implementation of the aria.jsunit.IRobot
interface to use the Selenium Java Robot.

The Selenium Java Robot tool is available here:
https://github.com/ariatemplates/selenium-java-robot

If the browser was launched with the Selenium Java Robot tool, it will
be detected automatically by the aria.jsunit.Robot class and the
corresponding implementation of the interface will be used.

close #1280