fix #1281 fixing random test failures with Slider and DoubleSlider
This commit makes aria.utils.dragdrop.Drag a static dependency of
the Slider and DoubleSlider widgets. It does not make sense to have
it as a dynamic dependency, as it is always needed, as soon as those
widgets are used. Moreover, using a dynamic dependency makes their test
fail randomly, as the dependency may not have been completely loaded
when the test starts, leading to errors.